@@ -1,6 +1,7 @@
 {
   "name": "nuxt-modules",
   "description": "Modules to supercharge Nuxt project",
+  "type": "module",
   "private": true,
   "scripts": {
     "preinstall": "npx only-allow pnpm",
@@ -9,7 +10,7 @@
     "stub": "pnpm -r --parallel stub",
     "changelogs": "pnpm -r --parallel changelog && pnpm prettier",
     "prettier": "prettier --config .prettierrc --write ./**/*.md",
-    "lint": "eslint . --ext .js,.ts --fix",
+    "lint": "eslint",
     "prepare": "husky install",
     "postinstall": "pnpm -r --parallel stub && cd packages/playground && pnpm nuxt-prepare"
   },
@@ -28,31 +29,29 @@
     "pnpm": ">=8.0.0"
   },
   "devDependencies": {
-    "@changesets/cli": "^2.27.1",
-    "@commitlint/cli": "^19.2.1",
-    "@commitlint/config-conventional": "^19.1.0",
-    "@types/node": "20.12.4",
-    "@typescript-eslint/eslint-plugin": "^7.5.0",
-    "@typescript-eslint/parser": "^7.5.0",
-    "c8": "^9.1.0",
-    "commitizen": "^4.3.0",
-    "conventional-changelog-angular": "^7.0.0",
-    "conventional-changelog-cli": "^4.1.0",
-    "cz-git": "^1.9.1",
-    "eslint": "^8.57.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "husky": "^9.0.11",
-    "lint-staged": "^15.2.2",
-    "prettier": "^3.2.5",
-    "typescript": "^5.4.3"
+    "@changesets/cli": "^2.27.12",
+    "@commitlint/cli": "^19.7.1",
+    "@commitlint/config-conventional": "^19.7.1",
+    "@noahyu/conventional-changelog-base": "^1.1.0",
+    "@noahyu/eslint-config-vue": "^1.4.2",
+    "@types/node": "22.13.0",
+    "c8": "^10.1.3",
+    "commitizen": "^4.3.1",
+    "conventional-changelog-angular": "^8.0.0",
+    "conventional-changelog-cli": "^5.0.0",
+    "cz-git": "^1.11.0",
+    "husky": "^9.1.7",
+    "lint-staged": "^15.4.3",
+    "prettier": "^3.4.2",
+    "typescript": "^5.7.3"
   },
   "lint-staged": {
-    "*.{js,ts}": "eslint --fix"
+    "*.{js,ts,jsx,tsx,vue}": "eslint --fix"
   },
   "config": {
     "commitizen": {
       "path": "./node_modules/cz-git"
     }
-  }
+  },
+  "packageManager": "pnpm@9.15.4"
 }
